details

+++ MiniProject.Business/CourseService/CourseService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using MiniProject.DataAccess.Data;
+using MiniProject.Models.Dtos;
 using MiniProject.Models.Models;
 
 namespace MiniProject.Business.CourseService;
@@ -13,8 +14,13 @@ public class CourseService : ICourseService
         _context = context;
     }
 
-    public async Task Create(Course course)
+    public async Task Create(CourseDto courseDto)
     {
+        var course = new Course
+        {
+            CourseName = courseDto.CourseName,
+        };
+
         _context.Courses.Add(course);
         _context.SaveChanges();
     }
@@ -22,24 +28,48 @@ public class CourseService : ICourseService
     public async Task Delete(int courseId)
     {
         var course = await _context.Courses.FindAsync(courseId);
+        if (course == null) return;
+
         _context.Courses.Remove(course);
         await _context.SaveChangesAsync();
     }
 
-    public async Task<List<Course>> GetAllCourses()
+    public async Task<List<CourseDto>> GetAllCourses()
     {
-        var courses = await _context.Courses.Include(s => s.Enrollments).ToListAsync();
+        var courses = await _context.Courses
+            .Select(c => new CourseDto
+            {
+                CourseId = c.CourseId,
+                CourseName = c.CourseName,
+                Unit = c.Unit,
+            }).ToListAsync();
         return courses;
     }
 
-    public async Task<Course> GetById(int id)
+    public async Task<CourseDto> GetById(int id)
     {
         var course = await _context.Courses.FindAsync(id);
-        return course;
+        if (course == null)
+        {
+            return null;
+        }
+
+        return new CourseDto
+        {
+            CourseId = course.CourseId,
+            CourseName = course.CourseName,
+            Unit = course.Unit
+        };
     }
 
-    public async Task Update(Course course)
+    public async Task Update(CourseDto courseDto)
     {
+        var course = await _context.Courses.FindAsync(courseDto.CourseId);
+        if (course == null) return;
+
+        course.CourseName = courseDto.CourseName;
+        course.Unit = courseDto.Unit;
+
         _context.Courses.Update(course);
         await _context.SaveChangesAsync();
     }
