+++ MiniProject.Business/TeacherService/TeacherService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using MiniProject.DataAccess.Data;
+using MiniProject.Models.Dtos;
 using MiniProject.Models.Models;
 
 namespace MiniProject.Business.TeacherService;
@@ -13,8 +14,14 @@ public class TeacherService : ITeacherService
         _context = context;
     }
 
-    public async Task Create(Teacher teacher)
+    public async Task Create(TeacherDto teacherDto)
     {
+        var teacher = new Teacher
+        {
+            FirstName = teacherDto.FirstName,
+            LastName = teacherDto.LastName,
+        };
+
         _context.Teachers.Add(teacher);
         _context.SaveChanges();
     }
@@ -22,24 +29,49 @@ public class TeacherService : ITeacherService
     public async Task Delete(int teacherId)
     {
         var teacher = await _context.Teachers.FindAsync(teacherId);
+        if (teacher == null) return;
+
         _context.Teachers.Remove(teacher);
         await _context.SaveChangesAsync();
     }
 
-    public async Task<List<Teacher>> GetAllTeachers()
+    public async Task<List<TeacherDto>> GetAllTeachers()
     {
-        var teachers = await _context.Teachers.Include(s => s.Enrollments).ToListAsync();
+        var teachers = await _context.Teachers
+            .Select(t => new TeacherDto
+            {
+                TeacherId = t.TeacherId,
+                FirstName = t.FirstName,
+                LastName = t.LastName
+            })
+            .ToListAsync();
         return teachers;
     }
 
-    public async Task<Teacher> GetById(int id)
+    public async Task<TeacherDto> GetById(int id)
     {
         var teacher = await _context.Teachers.FindAsync(id);
-        return teacher;
+        if (teacher == null)
+        {
+            return null;
+        }
+
+        return new TeacherDto
+        {
+            TeacherId = teacher.TeacherId,
+            FirstName = teacher.FirstName,
+            LastName = teacher.LastName
+        };
     }
 
-    public async Task Update(Teacher teacher)
+    public async Task Update(TeacherDto teacherDto)
     {
+        var teacher = await _context.Teachers.FindAsync(teacherDto.TeacherId);
+        if (teacher == null) return;
+
+        teacher.FirstName = teacherDto.FirstName;
+        teacher.LastName = teacherDto.LastName;
+
         _context.Teachers.Update(teacher);
         await _context.SaveChangesAsync();
     }
